+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Tidy README
 - Switch to eslint-config-airbnb-base
 - Drop support for node versions below 14
+- Migrate to zUnit
 
 ## 4.0.1